Add files via upload
This R code is initialized to take user input for the month, year, and type of pollutant in Hyderabad City between 2017 and 2022 and gives the user an overview of the levels of all the pollutants (in the specified month and year) as well as the specific pollutant (over the course of years 2017-2022) the user has entered into the program in the form of tables and graphs. The R code as well as the dataset needs to be downloaded in the same folder by the user to use the program.